Added Player and Spawns factories
