--- Syph.Core/Engine/Factories/PlayerFactory.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Syph.Core.Engine.Factories
-{
-    class PlayerFactory
-    {
-    }
-}
