--- Syph.Core/Engine/Factories/SpawnFactory.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Syph.Core.Engine
-{
-    public static class SpawnFactory
-    {
-        //TODO
-    }
-}
